+++ CHANGELOG.md
@@ -510,6 +510,13 @@
 
 ## 4.8.0
 
+- Added dynamic light sources culling ([#318](https://github.com/LambdAurora/LambDynamicLights/issues/318)).
+  - This should significantly reduce the amount of chunk section rebuilds processed, which has been one of the notorious sources of lag of this mod.
+  - You can disable culling in the Performance tab of the mod in case there's glitches.
+- Added adaptative ticking ([#319](https://github.com/LambdAurora/LambDynamicLights/issues/319)).
+  - Light sources far away will tick a bit slower, slowing more the further away they are.
+  - The distances (in chunks) for which light sources will begin to slow down is configurable in the new Performance tab of the mod.
+  - Light sources behind the camera will tick a lot slower.
 - Cleaned up a bit the settings screen and added explanations to the entities dynamic light sources tab.
 - Updated German translations ([#320](https://github.com/LambdAurora/LambDynamicLights/pull/320)).